feat(cache): move to redis package
